Made the amount editable through productBoxAmount-TextField.
Notable bugs: "Adding" 1 of a product to your shopping cart by clicking "-" (which is also a bug in the first place) will cause the text field to display the amount of that product in your cart -1. When loading a category, every textField will show 0, despite the true amount in cart.